Add POC for customizable API index page
