+++ plugins/api-docs/package.json
@@ -32,6 +32,7 @@
     "@asyncapi/react-component": "^0.19.2",
     "@backstage/catalog-model": "^0.7.5",
     "@backstage/core": "^0.7.4",
+    "@backstage/plugin-catalog": "^0.5.5",
     "@backstage/plugin-catalog-react": "^0.1.4",
     "@backstage/theme": "^0.2.5",
     "@material-icons/font": "^1.0.2",
