+++ plugins/api-docs/src/components/ApiExplorerPage/index.ts
@@ -15,4 +15,3 @@
  */
 
 export { ApiExplorerPage } from './ApiExplorerPage';
-export { ApiExplorerPageBase } from './ApiExplorerPageBase';
